pyproject.toml
[tool.poetry]
name = "helm-upgrade"
version = "0.1.0"
description = "A Python CLI to manage Helm Chart dependencies"
authors = ["Sarah Gibson <drsarahlgibson@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
bs4 = "0.0.1"
numpy = "1.22.2"
PyYAML = "6.0"
requests = "2.27.1"

[tool.poetry.dev-dependencies]
coverage = "6.3.1"
pytest = "7.0.0"
responses = "0.18.0"

[tool.poetry.scripts]
helm-upgrade = "helm_upgrade.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
